refactor downloadData to another file
