+++ src/components/data-management/ProjectDetails.jsx
@@ -23,7 +23,7 @@ import UploadDetailsModal from './UploadDetailsModal';
 import MetadataPopover from './MetadataPopover';
 
 import { trackAnalysisLaunched } from '../../utils/tracking';
-import { getFromApiExpectOK, getFromUrlExpectOK } from '../../utils/getDataExpectOK';
+import { getFromUrlExpectOK } from '../../utils/getDataExpectOK';
 import {
   deleteSamples, updateSample,
 } from '../../redux/actions/samples';
@@ -50,9 +50,8 @@ import '../../utils/css/hover.css';
 import runGem2s from '../../redux/actions/pipeline/runGem2s';
 
 import downloadTypes from '../../utils/downloadTypes';
-import pushNotificationMessage from '../../utils/pushNotificationMessage';
-import endUserMessages from '../../utils/endUserMessages';
 import pipelineStatusValues from '../../utils/pipelineStatusValues';
+import downloadData from '../../utils/downloadExperimentData';
 
 const { Title, Text, Paragraph } = Typography;
 
@@ -114,32 +113,6 @@ const ProjectDetails = ({ width, height }) => {
       setSampleNames(new Set());
     }
   }, [samples, activeProject]);
-
-  const downloadData = async (type) => {
-    const experimentId = activeProject?.experiments[0];
-
-    try {
-      if (!environment) throw new Error('Environment is not set');
-      if (!Object.values(downloadTypes).includes(type)) throw new Error('Invalid download type');
-
-      const { signedUrl } = await getFromApiExpectOK(`/v1/experiments/${experimentId}/download/${type}`);
-
-      const link = document.createElement('a');
-      link.style.display = 'none';
-      link.href = signedUrl;
-      link.download = `${type}_${experimentId}.rds`;
-
-      document.body.appendChild(link);
-      link.click();
-
-      setTimeout(() => {
-        URL.revokeObjectURL(link.href);
-        link.parentNode.removeChild(link);
-      }, 0);
-    } catch (e) {
-      pushNotificationMessage('error', endUserMessages.ERROR_DOWNLOADING_DATA);
-    }
-  };
 
   useEffect(() => {
     const { loading, error, status } = backendStatus;
@@ -647,7 +620,11 @@ const ProjectDetails = ({ width, height }) => {
               : 'Launch analysis to process data'
           }
           placement='left'
-          onClick={() => downloadData(downloadTypes.PROCESSED_SEURAT_OBJECT)}
+          onClick={() => {
+            // Change if we have more than one experiment per project
+            const experimentId = activeProject.experiments[0];
+            downloadData(experimentId, downloadTypes.PROCESSED_SEURAT_OBJECT);
+          }}
         >
           Processed Seurat object (.rds)
         </Tooltip>
